create dashboard factory
<!-- ps-id: 53dcf13c-4457-41de-bc9f-0c2d27d2f979 -->
