+++ packages/admin-config/src/dashboards/color.dashboard.ts
@@ -1,25 +1,26 @@
 import { Color } from '@repo/database';
 import { AdminFieldType } from '@repo/types';
-import { Dashboard, createDashboard } from '../dashboard';
+import { createDashboard } from '../create-dashboard';
 
-export function createColorDashboard(): Dashboard<Color> {
-  return createDashboard<Color>({
-    name: 'Color',
-    modelName: 'Color',
-    getDisplayName: (record: Color) => record.label,
-    attributeTypes: [
-      { name: 'label', type: AdminFieldType.STRING },
-      { name: 'hex', type: AdminFieldType.STRING },
-      { name: 'createdAt', type: AdminFieldType.DATETIME },
-      { name: 'updatedAt', type: AdminFieldType.DATETIME },
-    ],
-    collectionAttributes: ['label', 'hex', 'createdAt', 'updatedAt'],
-    collectionFilterAttributes: [],
-    showAttributes: ['label', 'hex', 'createdAt', 'updatedAt'],
-    createFormAttributes: ['label', 'hex'],
-    editFormAttributes: ['label', 'hex'],
+export const ColorDashboard = createDashboard<Color>({
+  name: 'Color',
+  modelName: 'Color',
 
-    // Text searchable attributes. Only supports String attribute types
-    searchAttributes: ['label'],
-  });
-}
+  getDisplayName: (record: Color) => record.label,
+
+  attributeTypes: [
+    { name: 'label', type: AdminFieldType.STRING },
+    { name: 'hex', type: AdminFieldType.STRING },
+    { name: 'createdAt', type: AdminFieldType.DATETIME },
+    { name: 'updatedAt', type: AdminFieldType.DATETIME },
+  ],
+
+  collectionAttributes: ['label', 'hex', 'createdAt', 'updatedAt'],
+  collectionFilterAttributes: [],
+  showAttributes: ['label', 'hex', 'createdAt', 'updatedAt'],
+  createFormAttributes: ['label', 'hex'],
+  editFormAttributes: ['label', 'hex'],
+
+  // Text searchable attributes. Only supports String attribute types
+  searchAttributes: ['label'],
+});
