+++ packages/admin-config/src/dashboards/product.dashboard.ts
@@ -1,27 +1,26 @@
-import { Product, User } from '@repo/database';
+import { Product } from '@repo/database';
 import { AdminFieldType } from '@repo/types';
-import { Dashboard, createDashboard } from '../dashboard';
+import { createDashboard } from '../create-dashboard';
 
-export function createProductDashboard(): Dashboard<Product> {
-  return createDashboard<Product>({
-    pinnedOnHome: true,
-    name: 'Product',
-    modelName: 'Product',
+export const ProductDashboard = createDashboard<Product>({
+  name: 'Product',
+  modelName: 'Product',
 
-    getDisplayName: (record: Product): string => record.name,
+  pinnedOnHome: true,
 
-    attributeTypes: [
-      { name: 'name', type: AdminFieldType.STRING },
-      { name: 'price', type: AdminFieldType.INTEGER },
-      { name: 'stock', type: AdminFieldType.INTEGER },
-    ],
-    collectionAttributes: ['name', 'price', 'stock'],
-    collectionFilterAttributes: ['name', 'price', 'stock'],
-    showAttributes: ['name', 'price', 'stock'],
-    createFormAttributes: ['name', 'price', 'stock'],
-    editFormAttributes: ['name', 'price', 'stock'],
+  getDisplayName: (record: Product): string => record.name,
 
-    // Text searchable attributes. Only supports String attribute types
-    searchAttributes: ['name'],
-  });
-}
+  attributeTypes: [
+    { name: 'name', type: AdminFieldType.STRING },
+    { name: 'price', type: AdminFieldType.INTEGER },
+    { name: 'stock', type: AdminFieldType.INTEGER },
+  ],
+  collectionAttributes: ['name', 'price', 'stock'],
+  collectionFilterAttributes: ['name', 'price', 'stock'],
+  showAttributes: ['name', 'price', 'stock'],
+  createFormAttributes: ['name', 'price', 'stock'],
+  editFormAttributes: ['name', 'price', 'stock'],
+
+  // Text searchable attributes. Only supports String attribute types
+  searchAttributes: ['name'],
+});
